backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-filter==23.2
django-polymorphic==3.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==6.6.0
Markdown==3.4.3
psycopg==3.1.9
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
zipp==3.15.0
